requirements.txt
ultralytics>=8.0.0
streamlit>=1.28.0
torch>=2.0.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
pyyaml>=6.0.0
tensorflowjs>=4.0.0
